requirements.txt
asgiref==3.8.1
Django==5.0.7
django-ckeditor==6.7.3
django-js-asset==3.1.2
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.7.0
